Move Dom<T> and Root<T> types to script_bindings.
